refactor: 클래스 및 필드명 개선, 패키지 구조 정리
- RefreshToken 클래스의 expiryDate 필드를 expiresAt으로 변경
- JwtUtil 클래스를 JwtProvider로 이름 변경
- CustomUserDetailsService를 global/security에서 domain/user 패키지로 이동
  - 사용자 도메인 정보에 의존하는 서비스이므로 관련 패키지로 이동
