+++ src/main/java/faithcoderlab/newdpraise/global/security/JwtProvider.java
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class JwtUtil {
+public class JwtProvider {
 
   private final JwtProperties jwtProperties;
